requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
boto3==1.34.74
botocore==1.34.74
certifi==2023.7.22
charset-normalizer==3.3.1
chromadb==0.3.22
click==8.1.7
clickhouse-connect==0.6.18
dataclasses-json==0.5.14
Django==4.2.6
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
duckdb==0.8.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.1
huggingface-hub==0.17.3
idna==3.4
incremental==21.3.0
InstructorEmbedding==1.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.240
langsmith==0.0.54
lit==16.0.6
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.25.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
posthog==3.0.2
psutil==5.9.6
ptyprocess==0.7.0
pyasn1==0.4.8
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==2.8.0
PyPika==0.48.9
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.10.1
safetensors==0.3.1
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
zstandard==0.21.0
